+++ doc/source/doxygen-docs/changelog.md
@@ -10,6 +10,8 @@
   - \ref mrpt_io_grp
     - GZIP compressed streams now also support open and append. See new mrpt::io::CFileGZOutputStream::open() signature.
     - New enum mrpt::io::OpenMode for clearer-to-read code.
+  - \ref mrpt_serialization_grp
+    - Implemented serialization of mrpt::containers::bimap in the new header `#include <mrpt/serialization/bimap_serialization.h>`.
   - \ref mrpt_system_grp
     - Backwards-compatible change: New function mrpt::system::InvalidTimeStamp() used now inside the macro INVALID_TIMESTAMP, so the macro always returns a const reference instead of returning by value.
 - BUG FIXES: